pyproject.toml
[project]
name = "hindsight"
version = "0.1.0"
description = "Hindsight — aggregate ActivityWatch + Claude Code + Codex CLI + other AI tool logs into a daily LLM-summarized journal. Export to Notion, Markdown, JSON."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "cab" }]
keywords = ["activitywatch", "claude-code", "codex", "notion", "journaling", "llm"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "anthropic>=0.40",
    "openai>=1.40",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5"]

[project.scripts]
hindsight = "hindsight.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hindsight"]

[tool.ruff]
line-length = 100
target-version = "py310"
